qcacmn: Add a duration check for dfs false detection
Various WIFI devices leak/send out calibration signal that is same
as radar signal. A duration check is added to filter them.

A new queue is created to collect radar pulse information of sidx(=1)
and sidx(=2) pulse. The queue can save 32 radar pulses. After radar is
found, for the radar pulses in the new queue, remove radar pulses that
happened 20millisecons ago. Then find the difference between the biggest
and the smallest duration of the rest radar pulses. If the difference
is more than 10microseconds, ignore this radar found.

Change-Id: I9bbb62b0d91946daccf2e176af578975cf968446
CRs-Fixed: 2227579