qcacmn: Fix memory leak in WMI
In case of invalid MIMO PS config in wmi cmd the wmi cmd
is not sent and we return. This leads to memory leak as wmi
net buf allocated is not freed before returning.

Fix the memory leak by freeing the wmi net buf in such case
before return.

Change-Id: I7f353f0b4d84a663801482a9c28296371bea3b89
CRs-fixed: 2230346