qcacmn: Fix the wlan can't work issue
After merging fixes for smmu and noc error issues, wlan can't work.
This change is used to fix the regression issue. The changes
includeing: a)change the force wake request implementation to the
previous patch of the mainline b)do not use the API
dp_ipa_tx_buf_smmu_mapping

Change-Id: Ic2d230ee88f750fbacad2ed93e993c006b43c42e