qcacmn: Do not return error if ol_ctx is NULL
For chips that do not use BMI, the epping enable process
will break if the epping_enable() flow is returned when
ol_ctx is NULL. This should not happen because ol_ctx
is only used when BMI is enabled. Whenever ol_ctx is referred
to, it will be checked for NULL before use, so we should
not return error here.

Change-Id: Ia873536a294b79f774193493646519163baf28c4
CRs-Fixed: 2901231