qca-wifi: Check DFS domain instead of configuration in ADFS SM
In Agile DFS SM, the RUNNING state event handler for the STOP event uses
the preCAC or RCAC configuration to identify which timer(RCAC/preCAC) to
abort and cleanup. Using the configuration, to make this decision will
not ensure abort or cleanup because it can be disabled at any time.

Consider the following scenario,

1) An AP has been brought up with preCAC enabled and the preCAC is in
INIT state and START event moves the preCAC to RUNNING state since there
was no preCAC that was already RUNNING.
2) An user space command is issued to disable the preCAC. The RUNNING state
receives a STOP event which will try to abort and clean up the preCAC.
Since abort and cleanup is based on the configuration which has now been
disabled it doesn't happen. However the STOP event forces a state change
to INIT.
3) When the user re-enables the preCAC, the START event will not be able to
move the preCAC to RUNNING state since the abort and cleanup has not taken
place.

The RUNNING state STOP event handler should have a DFS domian check instead
of preCAC or RCAC configuration check to decide which time to stop and
cleanup. When the domain is FCC do a RCAC abort and cleanup and when
the domain is ETSI do a preCAC abort and cleanup.

Change-Id: I03d4fb23eccd0f1a8535b831471ee83cbb7f9766
CRs-Fixed: 2995869