qcacmn: Rebuild reg chanlist when ch144 is enabled/disabled
Channel 144 is disabled by default in Japan country. Once we switch
country from Japan to US/ETSI using cfg80211tool, channel 144
is expected to be re-enabled. However, it is not re-enabled as
"en_chan_144" pdev config is not set once we switch from Japan to
other countries.

Set en_chan_144 to true if the country is not JAPAN in
"reg_modify_chan_list_for_japan" API.
Also, if "en_chan_144" is set to true, enable the channels in the
frequency range from 5710 to 5730 (inclusive of all the
5MHZ channels in-between) in reg_modify_chan_list_for_chan_144() and
if "en_chan_144" to false, disable all the channels in the
144HT20 frequency range including the overlapping frequencies.

CRs-Fixed: 3059729
Change-Id: Ie005413e945a2ecfaf774cbb969ade6b5c33d6f6