qcacmn: Replace nol entry cfreq with frequency ranges
With the introduction of 5/10MHz, the granularity of a DFS channel
can be 5/10/20MHz. Maintaining NOL entries as 20MHz centers will
result in 4-5 5MHz channels marked as NOL even though they're
unaffected.

Change the granularity of NOL entries to a minimum of 5MHz by using
channel frequency ranges as NOL data instead of cfreq.

Change-Id: Ibe6acf66b28f7a5f8ec79234019fa9604e8074bb