qcacmn: Fix the enable state for DFS channel
In regulatory component, for a non-DFS channel state 'ENABLE'
indicates it is enable. However, for a DFS channel, if the
channel is enabled, the state is not 'ENABLE' instead it is 'DFS'.

The function 'reg_dis_or_en_channel' while enabling a channel,
sets the state to 'ENABLE' even for a DFS channel.
Modify the function 'reg_dis_or_en_channel', such that for
a DFS channel it sets channel state to 'DFS' instead of 'ENABLE'.

Also, enable channel 144 only if both the regulatory rules and the device
capability support the channel 144, which is achieved
by searching the pdev reg rules and finding if the channel is supported
by chip ranges.

CRs-Fixed: 3077942
Change-Id: I073a28b93bd3a9a95bd0aa7e4dcbe6959c62588f