qcacmn: Remove duplicate scan chan entries
When a background scan is initiated on AP using "iw dev athX scan"
cmd with half/qtr rate flag set on a 5G radio, the scan gets aborted.
This is because scan timeout happens in FW. The number of scan
channel list sent to FW is 221. FW is unable to scan 221 channels
within the max scan timeout of 50000 msec.

In the 221 channels sent by host, there are duplicate entries of
same channel frequencies. This is because the scan channel entries are
populated by the cfg80211 layer and a set of 3 entries (one chan entry for
each 5/10/20MHZ) are sent to cfg80211 layer during cfg80211_init by the
host driver.

Hence filter out the duplicate scan channel entries sent to FW so that
scan timeout issue is solved.

Change-Id: I82acb3ee6bcba1d9db647b755eb95f76fe1a6f73
CRs-Fixed: 3075057