qcacmn: Do not add 6 GHz RNR flag in PNO scan
In suspend mode, driver issues the PNO scan with frequency list to
firmware. Currently, for 6 GHz PNO scan, driver adds the RNR flag
for non-PSC channels and sent the scan request to firmware. But
firmware is unable to scan for non-PSC frequencies as 2.4 GHz and
5 GHz frequencies are not present in the frequency list.

So, to fix this, driver exempts the addition of RNR flag for
non-PSC channel in PNO scan.

Change-Id: If2403586460f9da1647bc904027ed738e255af22
CRs-Fixed: 3384972