qcacmn: Consider vdev_id also along with link id for connected bitmap
When STA roams from SLO to SLO, only vdev 0 should be set as connected
link, while vdev 1 should not be set as connected link.

Currently vdev 1 is set as connected link incorrectly because its link
id is 0 which is the same with the link id from roam sync event.

link id is not a unique parameter. All are 0 by default.
vdev id should be used because it is a unique parameter.

Change-Id: I097b6f672c4656bfa2892f102e50998065463516
CRs-Fixed: 3497124