qcacmn: Handle invalid 6 GHz power type advertised by AP
If AP doesn't advertise 6 GHz power type or advertises
invalid 6 GHz power type, find the power type for connection
as per below logic.
a) Connect in VLP only if VLP power type is available for STA
   for the connection channel.
b) Otherwise connect in LPI only if LPI power type is available
   for STA for the connection channel for STA.
Don't connect if VLP or LPI power types are not available for
STA for the connection channel.

Change-Id: I4c2d146050bc49b856f22c15af36a2b1d229f448
CRs-Fixed: 3529927