qcacmn: Add check if client_type is available
The 'client_type' may not be assigned when reg_get_cur_6g_client_type()
is executed. Then 'client_type' is used for mas_chan_list_6g_client[][]
to set 'master_chan_list'. Because 'client_type' is a local variable
without initialization, it may cause unexpected behavior if not be set.

Change-Id: I061c9313f0d2767c8ffb30466fd2b69a3553a344