qcacmn: Align src ring shadow address only if necessary
Align the shadow_base by default will need extra allocated memory.
Although the extra memory itself is not large, kernel will
allocate large amount of useless memory if the useful memory size
is large because it's based on buddy algorithm. But most of the
time we do not need to align the memory because it's already
aligned by default.

Change-Id: I34fb07c620fcc3dc5a92cd07db777e8eec64afa1
CRs-Fixed: 3566377